requirements.txt
Flask
mysql-connector-python
enum34
jimit
requests
PyJWT
